changement ft_error

+++ srcs/utils/ft_error.c
@@ -12,7 +12,7 @@
 
 #include "utils.h"
 
-void	*ft_error_nb(char *str, char *str2, size_t line, size_t col)
+void	*ft_error_nb(char *str, char *str2, int line, int col)
 {
 	if (str2 == NULL)
 		ft_printf("Error [%d:%d]: %s\n", line, col, str);
